Move the fast control menu to corresponding TU
